+++ assigner.py
@@ -5,7 +5,6 @@ import logging
 import os
 import re
 import csv
-import tempfile
 
 from datetime import datetime
 
@@ -25,188 +24,16 @@ description = "An automated tool for assigning programming homework."
 
 subcommands = OrderedDict([
     ("new", "assigner.commands.new"),
+    ("assign", "assigner.commands.assign"),
+    ("open", "assigner.commands.open"),
+    ("get", "assigner.commands.get"),
+    ("lock", "assigner.commands.lock"),
+    ("unlock", "assigner.commands.unlock"),
+    ("archive", "assigner.commands.archive"),
+    ("unarchive", "assigner.commands.unarchive"),
+    ("status", "assigner.commands.status"),
+    ("import", "assigner.commands.import"),
 ])
-
-
-@config_context
-def assign(conf, args):
-    """Creates homework repositories for an assignment for each student
-    in the roster.
-    """
-    hw_name = args.name
-    if args.branch:
-        branch = args.branch
-    else:
-        branch = "master"
-    dry_run = args.dry_run
-    force = args.force
-    host = conf.gitlab_host
-    namespace = conf.namespace
-    token = conf.token
-    semester = conf.semester
-
-    roster = get_filtered_roster(conf.roster, args.section, args.student)
-
-    actual_count = 0  # Represents the number of repos actually pushed to
-    student_count = len(roster)
-
-    with tempfile.TemporaryDirectory() as tmpdirname:
-        print("Assigning '{}' to {} student{} in {}.".format(
-            hw_name, student_count,
-            "s" if student_count != 1 else "",
-            "section " + args.section if args.section else "all sections")
-        )
-        base = BaseRepo(host, namespace, hw_name, token)
-        if not dry_run:
-            base.clone_to(tmpdirname, branch)
-        if force:
-            logging.warning("Repos will be overwritten.")
-        for i, student in enumerate(roster):
-            username = student["username"]
-            student_section = student["section"]
-            full_name = StudentRepo.name(semester, student_section,
-                                         hw_name, username)
-            repo = StudentRepo(host, namespace, full_name, token)
-
-            print("{}/{} - {}".format(i+1, student_count, full_name))
-            if not repo.already_exists():
-                if not dry_run:
-                    repo = StudentRepo.new(base, semester, student_section,
-                                           username, token)
-                    repo.push(base, branch)
-                actual_count += 1
-                logging.debug("Assigned.")
-            elif force:
-                logging.info("{}: Already exists.".format(full_name))
-                logging.info("{}: Deleting...".format(full_name))
-                if not dry_run:
-                    repo.delete()
-                    # HACK: Gitlab will throw a 400 if you delete and immediately recreate a repo.
-                    # A bit more than half a second was experimentally determined to prevent this issue.
-                    time.sleep(0.55)
-                    repo = StudentRepo.new(base, semester, student_section,
-                                           username, token)
-                    repo.push(base, branch)
-                actual_count += 1
-                logging.debug("Assigned.")
-            elif args.branch:
-                logging.info("{}: Already exists.".format(full_name))
-                # If we have an explicit branch, push anyways
-                repo.push(base, branch) if not dry_run else None
-                actual_count += 1
-                logging.debug("Assigned.")
-            else:
-                logging.warning("Skipping...")
-            i += 1
-
-    print("Assigned '{}' to {} student{}.".format(
-        hw_name,
-        actual_count,
-        "s" if actual_count != 1 else ""
-    ))
-    if actual_count == 0:
-        logging.warning(
-            "Consider using --force if you want to override existing repos."
-        )
-
-
-@config_context
-def open_assignment(conf, args):
-    """Adds each student in the roster to their respective homework
-    repositories as Developers so they can pull/commit/push their work.
-    """
-    hw_name = args.name
-    host = conf.gitlab_host
-    namespace = conf.namespace
-    token = conf.token
-    semester = conf.semester
-
-    roster = get_filtered_roster(conf.roster, args.section, args.student)
-
-    count = 0
-    for student in roster:
-        username = student["username"]
-        student_section = student["section"]
-        full_name = StudentRepo.name(semester, student_section,
-                                     hw_name, username)
-
-        try:
-            repo = StudentRepo(host, namespace, full_name, token)
-            if "id" not in student:
-                student["id"] = Repo.get_user_id(username, host, token)
-
-            repo.add_member(student["id"], Access.developer)
-            count += 1
-        except RepoError:
-            logging.warn("Could not add {} to {}.".format(username, full_name))
-        except HTTPError:
-            raise
-
-    print("Granted access to {} repositories.".format(count))
-
-
-@config_context
-def get(conf, args):
-    """Creates a folder for the assignment in the CWD (or <path>, if specified)
-    and clones each students' repository into subfolders.
-    """
-    hw_name = args.name
-    hw_path = args.path
-    host = conf.gitlab_host
-    namespace = conf.namespace
-    token = conf.token
-    semester = conf.semester
-
-    roster = get_filtered_roster(conf.roster, args.section, args.student)
-
-    path = os.path.join(hw_path, hw_name)
-    os.makedirs(path, mode=0o700, exist_ok=True)
-
-    count = 0
-    for student in roster:
-        username = student["username"]
-        student_section = student["section"]
-        full_name = StudentRepo.name(semester, student_section,
-                                     hw_name, username)
-
-        try:
-            repo = StudentRepo(host, namespace, full_name, token)
-            repo.clone_to(os.path.join(path, username))
-            count += 1
-        except RepoError as e:
-            logging.warn(str(e))
-        except HTTPError as e:
-            if e.response.status_code == 404:
-                logging.warn("Repository {} does not exist.".format(full_name))
-            else:
-                raise
-
-    print("Cloned {} repositories.".format(count))
-
-
-def lock(args):
-    """Sets each student to Reporter status on their homework repository so
-    they cannot push changes, etc.
-    """
-    return manage_users(args, Access.reporter)
-
-
-def unlock(args):
-    """Sets each student to Developer status on their homework repository.
-    """
-    return manage_users(args, Access.developer)
-
-
-def archive(args):
-    """Archive each student repository so it won't show up in the project list.
-    """
-    return manage_repos(args, 'archive')
-
-
-def unarchive(args):
-    """Unarchive each student repository so it will show back up in the project list.
-    """
-    return manage_repos(args, 'unarchive')
 
 
 @config_context
@@ -247,131 +74,6 @@ def manage_users(conf, args, level):
             raise
 
     print("Changed {} repositories.".format(count))
-
-
-@config_context
-def status(conf, args):
-    """Retrieves and prints the status of repos"""
-    hw_name = args.name
-
-    if not hw_name:
-        raise ValueError("Missing assignment name.")
-
-    host = conf.gitlab_host
-    namespace = conf.namespace
-    token = conf.token
-    semester = conf.semester
-
-    roster = get_filtered_roster(conf.roster, args.section, args.student)
-    sort_key = args.sort
-
-    if sort_key:
-        roster.sort(key=lambda s: s[sort_key])
-
-    output = PrettyTable([
-        "#", "Sec", "SID", "Name", "Status", "Branches",
-        "HEAD", "Last Commit Author", "Last Commit Time"])
-    output.align["Name"] = "l"
-    output.align["Last Commit Author"] = "l"
-
-    progress = ProgressBar(max_value=len(roster))
-
-    for i, student in enumerate(roster):
-        progress.update(i)
-
-        name = student["name"]
-        username = student["username"]
-        student_section = student["section"]
-        full_name = StudentRepo.name(semester, student_section,
-                                     hw_name, username)
-
-        row = [i+1, student_section, username, name, "", "", "", "", ""]
-
-        try:
-            repo = StudentRepo(host, namespace, full_name, token)
-
-            if not repo.already_exists():
-                row[4] = "Not Assigned"
-                output.add_row(row)
-                continue
-
-            if "id" not in student:
-                student["id"] = Repo.get_user_id(username, host, token)
-
-            members = repo.list_members()
-            if student["id"] not in [s["id"] for s in members]:
-                row[4] = "Not Opened"
-                output.add_row(row)
-                continue
-
-            if repo.info["archived"]:
-                row[4] = 'Archived'
-            else:
-                level = Access([s["access_level"] for s in members if s["id"] == student["id"]][0])
-                row[4] = "Open" if level is Access.developer else "Locked"
-
-            branches = repo.list_branches()
-
-            if branches:
-                row[5] = ", ".join([b["name"] for b in branches])
-
-            commits = repo.list_commits()
-
-            if commits:
-                head = commits[0]
-                row[6] = head["short_id"]
-                row[7] = head["author_name"]
-                created_at = head["created_at"]
-                # Fix UTC offset format in GitLab's datetime
-                created_at = created_at[:-6] + created_at[-6:].replace(':', '')
-                row[8] = datetime.strptime(created_at, "%Y-%m-%dT%H:%M:%S.%f%z")
-
-            output.add_row(row)
-
-        except HTTPError:
-            raise
-
-    progress.finish()
-    print(output)
-
-
-@config_context
-def import_students(conf, args):
-    """Imports students from a CSV file to the roster.
-    """
-    section = args.section
-
-    # TODO: This should probably move to another file
-    email_re = re.compile(r"^(?P<user>[^@]+)")
-    with open(args.file) as fh:
-        reader = csv.reader(fh)
-
-        if "roster" not in conf:
-            conf["roster"] = []
-
-        # Note: This is incredibly hardcoded.
-        # However, peoplesoft never updates anything, so we're probably good.
-        reader.__next__()  # Skip the header
-        count = 0
-        for row in reader:
-            count += 1
-            match = email_re.match(row[4])
-            conf.roster.append({
-                "name": row[3],
-                "username": match.group("user"),
-                "section": section
-            })
-
-            try:
-                conf.roster[-1]["id"] = Repo.get_user_id(
-                    match.group("user"), conf.gitlab_host, conf.token
-                )
-            except RepoError:
-                logger.warning(
-                    "Student {} does not have a Gitlab account.".format(row[3])
-                )
-
-    print("Imported {} students.".format(count))
 
 
 @config_context
@@ -550,124 +252,6 @@ def make_parser():
         module = importlib.import_module(path)
         subparser = subparsers.add_parser(name, help=module.help)
         module.setup_parser(subparser)
-
-    # "assign" command
-    subparser = subparsers.add_parser("assign",
-                                      help="Assign a base repo to students")
-    subparser.add_argument("name",
-                           help="Name of the assignment to assign.")
-    subparser.add_argument("--branch", nargs="?",
-                           help="Branch to push")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to assign homework to")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of the student to assign to.")
-    subparser.add_argument("--dry-run", action="store_true",
-                           help="Don't actually do it.")
-    subparser.add_argument("-f", "--force", action="store_true", dest="force",
-                           help="Delete and recreate already existing " +
-                                "student repos.")
-    subparser.set_defaults(run=assign)
-
-    # "open" command
-    subparser = subparsers.add_parser("open", help="Grant students access " +
-                                                   "to their repos")
-    subparser.add_argument("name", help="Name of the assignment to grant " +
-                                        "access to")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to grant access to")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of the student to assign to.")
-    subparser.set_defaults(run=open_assignment)
-
-    # "get" command
-    subparser = subparsers.add_parser("get",
-                                      help="Clone student repos")
-    subparser.add_argument("name",
-                           help="Name of the assignment to retrieve.")
-    subparser.add_argument("path", default=".", nargs="?",
-                           help="Path to clone student repositories to")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to retrieve")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of student whose assignment needs " +
-                                "retrieving.")
-    subparser.set_defaults(run=get)
-
-    # "lock" command
-    subparser = subparsers.add_parser("lock",
-                                      help="Lock students out of repos")
-    subparser.add_argument("name",
-                           help="Name of the assignment to lock.")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to lock")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of student whose assignment needs " +
-                                "locking.")
-    subparser.add_argument("--dry-run", action="store_true",
-                           help="Don't actually do it.")
-    subparser.set_defaults(run=lock)
-
-    # "unlock" command
-    subparser = subparsers.add_parser("unlock",
-                                      help="unlock students from repos")
-    subparser.add_argument("name",
-                           help="Name of the assignment to unlock.")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to unlock")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of student whose assignment needs " +
-                                "unlocking.")
-    subparser.add_argument("--dry-run", action="store_true",
-                           help="Don't actually do it.")
-    subparser.set_defaults(run=unlock)
-
-    # "archive" command
-    subparser = subparsers.add_parser("archive",
-                                      help="Archive students repos")
-    subparser.add_argument("name",
-                           help="Name of the assignment to archive.")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to archive")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of student whose assignment to archive.")
-    subparser.add_argument("--dry-run", action="store_true",
-                           help="Don't actually do it.")
-    subparser.set_defaults(run=archive)
-
-    # "unarchive" command
-    subparser = subparsers.add_parser("unarchive",
-                                      help="Unarchive students repos")
-    subparser.add_argument("name",
-                           help="Name of the assignment to unarchive.")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to unarchive")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of student whose assignment to unarchive.")
-    subparser.add_argument("--dry-run", action="store_true",
-                           help="Don't actually do it.")
-    subparser.set_defaults(run=unarchive)
-
-    # "status" command
-    subparser = subparsers.add_parser("status",
-                                      help="Retrieve status of repos")
-    subparser.add_argument("--section", nargs="?",
-                           help="Section to get status of")
-    subparser.add_argument("--student", metavar="id",
-                           help="ID of student.")
-    subparser.add_argument("--sort", nargs="?", default="name",
-                           choices=["name", "username"],
-                           help="Key to sort users by.")
-    subparser.add_argument("name", nargs="?",
-                           help="Name of the assignment to look up.")
-    subparser.set_defaults(run=status)
-
-    # "import" command
-    subparser = subparsers.add_parser("import",
-                                      help="Import students from a csv")
-    subparser.add_argument("file", help="CSV file to import from")
-    subparser.add_argument("section", help="Section being imported")
-    subparser.set_defaults(run=import_students)
 
     # "canvas_import" command
     subparser = subparsers.add_parser("canvas_import",
